+++ src/backends/rustler/trait_bridge.rs
@@ -20,6 +20,7 @@ mod bridge_functions;
 mod generator;
 mod json_args;
 mod methods;
+mod native_args;
 #[cfg(test)]
 mod tests;
 mod visitor_bridge;